old/vpamnmun_dag/requirements.txt
# To install all dependencies, run: pip install -r requirements.txt
requests
oracledb
python-dotenv
tqdm
python-dateutil